Added console cursor manager. Added cuonsole handle getter
Added files: console.h, console.c, console_cursor.h, console_cursor.c. console file is responsible for getting the handle to the console which can be used later to e.g change console cursor visibility. console_cursor is responsible for moving the console cursor and for hiding it and showing it.